refactor: extract config JSON helpers
